requirements.txt
streamlit
pandas==1.3.5
numpy
pillow
xgboost==1.3.3
openpyxl
